back/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16.1,<3.17
gunicorn>=22.0,<23
whitenoise>=6.7,<7
psycopg2-binary>=2.9.9,<2.10
requests>=2.32.5,<2.33
django-environ>=0.12.0,<0.13
opencv-python-headless>=4.12.0.88,<4.13
numpy>=2.0.2,<2.1
scikit-learn>=1.6.1,<1.7
groq>=0.36.0,<0.37
<<<<<<< Updated upstream:back/requirements.txt
django-cors-headers>=4.9.0,<5.0
=======
<<<<<<< Updated upstream:requirements.txt
=======
django-cors-headers>=4.9.0,<5.0
django-allauth>=65.14.0,<66.0
PyJWT>=2.11.0,<2.12
cryptography>=46.0.4,<47.0
pillow>=12.1.0,<12.2.0
>>>>>>> Stashed changes:back/requirements.txt
>>>>>>> Stashed changes:requirements.txt
